Create basic setup
Create an images directory
and populate it with 3
images that will be used
as slides.

Setup the basic markup
in index.html.
A slideshow-container
contains the picture
frame (which will display
the current picture)
and the
slide images container
which will the carousel's
images.

In the CSS, add basic styles.
Style the hide class
display: 0 so that the
slide images container
and its images will be
hidden until needed.