Initialize ImageCarousel
Create an
image-carousel-module and
its respective file.

Initialize an ImageCarousel
class. In its constructor
select the nodelist of
all images and convert the
nodelist to an imagesArray
array.

Select the picture frame
and call the initialize
method if the
imagesArray has more
than one element.

For now, the initialize()
method just appends the
first image to the
picture frame.

In the index.js, create
an instance of
ImageCarousel class
when the whole page has
loaded.