Handle next button click
In the constructor, initialize
an counter index property and set
it to 1.

Initialize a nextSlide method()
that checks if index is
less than the length of the
images array (this is because
the first image is already
displayed onload and since index
starts at 1, it should display
elements and 1 and 2, supposing
there are 3 images in total),
it clears the picture frame,
appends the element that
corresponds to the index of
this.index in the imagesArray
and increments this.index.

Initialize a handleClicks()
method that selects the
next button and attaches a
click event listener to it.
When the next button is clicked,
the nextSlide() method is called.

Call handleClicks() in intialize()
so that when the class initializes,
the event listener attached to
the next button is set.

This setup displays the next
image on click of the next button.