Reappend images to images container
In the previous commit, the clearing
the picture frame of the previous
image before appending the next images
was ensuring that the previous image
was unappended from the
slide images container.

In the constructor, initialize
a oldImageIndex property
and set it to 0.

To prevent this, initialize a
reappendImageToImagesContainer()
method that takes in an
imageIndex arg. The method selects
the slide images container and
appends the image with the index
of imageIndex in the imagesArray
to it.

Call this method in nextSlide()
with oldImageIndex as its arg
and increment oldImageIndex
after. oldImageIndex was set to
0 so that the
reappendImageToImagesContainer()
can first reappend the first image  in
the imagesArray to the
slide images container before
appending others.