Handle nextSlide() better
In the constructor, change the
value of index to 0 and add 1 to
it in nextSlide(). This is because
the image with the index of 0
is already being rendered in the
picture frame. Now everytime that
nextSlide() is called, it adds 1
to the current index befor it does
the check in the if block.

Initialize a resetIndex() method
that resets the index to
imagesArray.length - 1.
The reset is necessary because
if the nextButton is clicked
several times after the
the last image in the imageArray
has been rendered, it increases
the value of the index to
a value greater than
imagesArray.length - 1.
Clicking on a previous button
(if the same index was used)
will try to render an image
with an index that is not
available in the imagesArray.

Call the resetIndex() after
the IF block in nextSlide so
that index is reset in case
it ever gets greater than
imagesArray.length - 1.