Handle previousSlide()
Initialize a previousSlide()
method that checks if index
is not equal to 0. If true, it
clears the pictureFrame,
decrements the index and
and renders the picture that
hold the position of the
current index in the
imagesArray.

In handleClicks(), select the
previousButton and attach
a click event listener to it
so that the previousSlide()
method is called on click.