Add basic styles
In the index.css,
style the body,
slideshow-container,
slide, images,
navigation buttons
slide-number,
image caption and the dot.

Give slide a display of
none so that all the
"slides" are hidden and
later reavealed with
javascript.

In the index.js, change
one classnames of the
images from slide to image
so that they can be
targeted by CSS.