Change the setup
In the CSS, setup the
slideshow so that
the slides are lined
up horizontally in the
slideshow container.
The slides are arranged
horizontally
slideshow container
with flex and
flex-shink of 0
on the slides make
each slide to take the
full width of the
slideshow container
(whenever it is visible)

Style the slideshow
container with an
overflow: hidden so that
only one slide is visible
at a time.

This setup prepares to have
the slides
transform: translate() to
a certain value to
the left so that the
previous slide moves a 100%
of its width to the left
to reveal the next or previous
slide as the case may be.