Reset currentSlideIndex
Introduce two methods -
resetCurrentIndexForNextSlide()
checks if the current slide index
is equal to the length of the
images array - 1 and if it is,
it resets it to 0.

resetCurrentIndexForPreviousSlide()
checks if the current index is
equal to 0, and if it is, it
resets it to 3.

Call these methods inside
nextSlide() and previousSlide()
respectively.

This setup ensures that the
carousel 'rotates' - when the
next button navigates to the
last image, it returns to
the first image (ready
to nevigate to the second image
on the next click) and the
previous button navigates to
the last image (ready
to nevigate to the second
image, if it is clicked again)
when the
first element was
previously in view. A rotating
setup.