Implement dot advancement for nextSlide
Select the nodelist of allDots.

Define a method
removeActiveClassFromAllDots() that
removes the active class from
all dots. This is necessary so
that all dots will be cleared
from any previous ative class
before a new
active class is added to
the current dot.

Define another method that
adds the active class to the
dot that is equivalent to the
slide that currently in the
viewport for when the nextSlide
method is called. It loops
through allDots and adds the
active class to the dot that
is the index of
currentIndex + 1. The way the
current index is incremented
here is the same way it is
done in nextSlide().

Call these 2 methods
in nextSlide.

This setup ensures that when
as advancement is made to
the next slide, advancement
is also 'seen' in an
equivalent dot.