Implement dot navigation
First use the spread operator
to unpack properties from the
allDots nodelist.

Define a method jumpToSlide
that takes in
a slideIndex as an arg.
It loops through all the slides
and transfroms each of them
to the set of values that will
bring the slide, whose
index(in allSlides) was
passed into it, to the viewport.

Consider this:

Slide Two, has an index of 1 in
allSlides and requires
transfrom values of

`Slide 1 : translate(-100%)`
`Slide 2: translate(-100%)`
`Slide 2: translate(-100%)`

to be visible in the viewport.
Assign the value of slideIndex
to currentSlideIndex so that
the currentSlideIndex stays
ready for nextButton and
previousButton clicks after
jumpToSlide.

In handleClicks, add click
event listeners to each
dot and call the jumpToSlide
with the respective arg
for each.

With this setup, clicking
on any dot navigates to
the respective slide.