Add active class to dots by index
Set up a method
addActiveClassToDotByIndex(dotIndex)
that that takes
in a dotIndex and adds the
active class to the dot whose
index in the allDots nodelist
matches the dotIndex in the
argument. This is so that a
dotIndex can be passed
and the active class
added to any dot as needed.

In the event listeners of all
the dots, call the
removeActiveClassFromAllDots
first so that all previous
active classes can be removed
and then call the
addActiveClassToDotByIndex()
with the respective args.

This setup ensures that
when any dot is clicked on,
only it, has the active class.