Advance slides after five seconds
Initialize a method that uses
a recursive setTimeout call to
create a self-perpetuating cycle
of advancing the slides every
five seconds.

When the method is first called,
it sets a timeout to 5 seconds.
After 5 seconds, nextSlide is
called and then it calls itself
recursively. This recursive call
sets another 5 second timeout
and the cycle continues.

Call the method in initialize
so that the method is called
when the class is first
initialized.